mb_memberboard 230518 page max board complite

+++ src/main/java/com/icia/mbboard/service/CommentService.java
@@ -28,7 +28,7 @@ public class CommentService {
         commentRepository.commentDelete(id);
     }
 
-    public Integer findCommentCntByBoardId(Long id) {
+    public Long findCommentCntByBoardId(Long id) {
         return commentRepository.findCommentCntByboardId(id);
     }
 }
